Move SDK private NoopTextMapPropagator directly into Configurator
