+++ sdk/lib/opentelemetry/sdk.rb
@@ -74,7 +74,6 @@ module OpenTelemetry
 end
 
 require 'opentelemetry/sdk/configurator'
-require 'opentelemetry/sdk/context'
 require 'opentelemetry/sdk/forwarding_logger'
 require 'opentelemetry/sdk/internal'
 require 'opentelemetry/sdk/instrumentation_library'
