--- sdk/lib/opentelemetry/sdk/context.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-# Copyright The OpenTelemetry Authors
-#
-# SPDX-License-Identifier: Apache-2.0
-
-module OpenTelemetry
-  module SDK
-    class Context < OpenTelemetry::Context
-    end
-  end
-end
-
-require 'opentelemetry/sdk/context/propagation'
